+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/models/DocumentCaption.java
@@ -40,7 +40,7 @@ public final class DocumentCaption {
      *
      * @param content the content value to set.
      */
-    void setContent(String content) {
+    private void setContent(String content) {
         this.content = content;
     }
 
@@ -58,7 +58,7 @@ public final class DocumentCaption {
      *
      * @param boundingRegions the boundingRegions value to set.
      */
-    void setBoundingRegions(List<BoundingRegion> boundingRegions) {
+    private void setBoundingRegions(List<BoundingRegion> boundingRegions) {
         this.boundingRegions = boundingRegions;
     }
 
@@ -76,7 +76,7 @@ public final class DocumentCaption {
      *
      * @param spans the spans value to set.
      */
-    void setSpans(List<DocumentSpan> spans) {
+    private void setSpans(List<DocumentSpan> spans) {
         this.spans = spans;
     }