+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/models/DocumentWord.java
@@ -49,7 +49,7 @@ public final class DocumentWord {
      * @param content the content value to set.
      * @return the DocumentWord object itself.
      */
-    void setContent(String content) {
+    private void setContent(String content) {
         this.content = content;
     }
 
@@ -72,7 +72,7 @@ public final class DocumentWord {
      * @param boundingPolygon the boundingPolygon value to set.
      * @return the DocumentWord object itself.
      */
-    void setBoundingPolygon(List<Point> boundingPolygon) {
+    private void setBoundingPolygon(List<Point> boundingPolygon) {
         this.boundingPolygon = boundingPolygon;
     }
 
@@ -91,7 +91,7 @@ public final class DocumentWord {
      * @param span the span value to set.
      * @return the DocumentWord object itself.
      */
-    void setSpan(DocumentSpan span) {
+    private void setSpan(DocumentSpan span) {
         this.span = span;
     }
 
@@ -110,7 +110,7 @@ public final class DocumentWord {
      * @param confidence the confidence value to set.
      * @return the DocumentWord object itself.
      */
-    void setConfidence(float confidence) {
+    private void setConfidence(float confidence) {
         this.confidence = confidence;
     }