+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/AnalyzeLayout.java
@@ -8,6 +8,7 @@ import com.azure.ai.formrecognizer.documentanalysis.DocumentAnalysisClientBuilde
 import com.azure.ai.formrecognizer.documentanalysis.models.AnalyzeResult;
 import com.azure.ai.formrecognizer.documentanalysis.models.DocumentOperationResult;
 import com.azure.ai.formrecognizer.documentanalysis.models.DocumentTable;
+import com.azure.ai.formrecognizer.documentanalysis.models.Point;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.polling.SyncPoller;
@@ -16,6 +17,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Sample for analyzing layout information from a document given through a file.
@@ -54,9 +56,9 @@ public class AnalyzeLayout {
 
             // lines
             documentPage.getLines().forEach(documentLine ->
-                System.out.printf("Line '%s; is within a bounding box %s.%n",
+                System.out.printf("Line '%s; is within a bounding polygon %s.%n",
                     documentLine.getContent(),
-                    documentLine.getBoundingPolygon().toString()));
+                    getBoundingCoordinates(documentLine.getBoundingPolygon())));
 
             // words
             documentPage.getWords().forEach(documentWord ->
@@ -66,9 +68,9 @@ public class AnalyzeLayout {
 
             // selection marks
             documentPage.getSelectionMarks().forEach(documentSelectionMark ->
-                System.out.printf("Selection mark is '%s' and is within a bounding box %s with confidence %.2f.%n",
+                System.out.printf("Selection mark is '%s' and is within a bounding polygon %s with confidence %.2f.%n",
                     documentSelectionMark.getSelectionMarkState().toString(),
-                    documentSelectionMark.getBoundingPolygon().toString(),
+                    getBoundingCoordinates(documentSelectionMark.getBoundingPolygon()),
                     documentSelectionMark.getConfidence()));
         });
 
@@ -88,5 +90,13 @@ public class AnalyzeLayout {
         // styles
         analyzeLayoutResult.getStyles().forEach(documentStyle
             -> System.out.printf("Document is handwritten %s%n.", documentStyle.isHandwritten()));
+    }
+
+    /**
+     * Utility function to get the bounding polygon coordinates.
+     */
+    private static String getBoundingCoordinates(List<Point> boundingPolygon) {
+        return boundingPolygon.stream().map(point -> String.format("[%.2f, %.2f]", point.getX(),
+            point.getY())).collect(Collectors.joining(", "));
     }
 }