+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -4,6 +4,7 @@
 This version of the SDK defaults to the latest supported API version, which currently is 2021-09-30-preview.
 
 ### Features Added
+- Added a method `getWords()` on model `DocumentLine`.
 
 ### Breaking Changes
 - Removed property `kind` from model `DocumentPage`